requirements.txt
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
executing==2.1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
ipython==8.27.0
jedi==0.19.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
prompt_toolkit==3.0.47
pure_eval==0.2.3
Pygments==2.18.0
python-telegram-bot==21.5
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
whitenoise==6.7.0
pillow